Standardized layout of vectors and moved all logic into cpp files
