+++ core/math/vector2.cpp
@@ -34,14 +34,7 @@
 
 #include "core/math/vector2i.h"
 #include "core/string/ustring.h"
-
-real_t Vector2::angle() const {
-	return Math::atan2(y, x);
-}
-
-Vector2 Vector2::from_angle(real_t p_angle) {
-	return Vector2(Math::cos(p_angle), Math::sin(p_angle));
-}
+#include "core/typedefs.h"
 
 real_t Vector2::length() const {
 	return Math::sqrt(x * x + y * y);
@@ -50,6 +43,44 @@ real_t Vector2::length() const {
 real_t Vector2::length_squared() const {
 	return x * x + y * y;
 }
+
+Vector2 Vector2::limit_length(real_t p_len) const {
+	const real_t l = length();
+	Vector2 v = *this;
+	if (l > 0 && p_len < l) {
+		v /= l;
+		v *= p_len;
+	}
+
+	return v;
+}
+
+
+Vector2 Vector2::sign() const {
+	return Vector2(SIGN(x), SIGN(y));
+}
+
+_FORCE_INLINE_ Vector2 Vector2::abs() const {
+	return Vector2(Math::abs(x), Math::abs(y));
+}
+
+Vector2 Vector2::floor() const {
+	return Vector2(Math::floor(x), Math::floor(y));
+}
+
+Vector2 Vector2::ceil() const {
+	return Vector2(Math::ceil(x), Math::ceil(y));
+}
+
+Vector2 Vector2::round() const {
+	return Vector2(Math::round(x), Math::round(y));
+}
+
+bool Vector2::is_finite() const {
+	return Math::is_finite(x) && Math::is_finite(y);
+}
+
+
 
 void Vector2::normalize() {
 	real_t l = x * x + y * y;
@@ -71,44 +102,9 @@ bool Vector2::is_normalized() const {
 	return Math::is_equal_approx(length_squared(), 1, (real_t)UNIT_EPSILON);
 }
 
-real_t Vector2::distance_to(const Vector2 &p_vector2) const {
-	return Math::sqrt((x - p_vector2.x) * (x - p_vector2.x) + (y - p_vector2.y) * (y - p_vector2.y));
-}
 
-real_t Vector2::distance_squared_to(const Vector2 &p_vector2) const {
-	return (x - p_vector2.x) * (x - p_vector2.x) + (y - p_vector2.y) * (y - p_vector2.y);
-}
-
-real_t Vector2::angle_to(const Vector2 &p_vector2) const {
-	return Math::atan2(cross(p_vector2), dot(p_vector2));
-}
-
-real_t Vector2::angle_to_point(const Vector2 &p_vector2) const {
-	return (p_vector2 - *this).angle();
-}
-
-real_t Vector2::dot(const Vector2 &p_other) const {
-	return x * p_other.x + y * p_other.y;
-}
-
-real_t Vector2::cross(const Vector2 &p_other) const {
-	return x * p_other.y - y * p_other.x;
-}
-
-Vector2 Vector2::sign() const {
-	return Vector2(SIGN(x), SIGN(y));
-}
-
-Vector2 Vector2::floor() const {
-	return Vector2(Math::floor(x), Math::floor(y));
-}
-
-Vector2 Vector2::ceil() const {
-	return Vector2(Math::ceil(x), Math::ceil(y));
-}
-
-Vector2 Vector2::round() const {
-	return Vector2(Math::round(x), Math::round(y));
+real_t Vector2::angle() const {
+	return Math::atan2(y, x);
 }
 
 Vector2 Vector2::rotated(real_t p_by) const {
@@ -119,59 +115,112 @@ Vector2 Vector2::rotated(real_t p_by) const {
 			x * sine + y * cosi);
 }
 
-Vector2 Vector2::posmod(real_t p_mod) const {
-	return Vector2(Math::fposmod(x, p_mod), Math::fposmod(y, p_mod));
+real_t Vector2::angle_to(const Vector2 &p_vector2) const {
+	return Math::atan2(cross(p_vector2), dot(p_vector2));
 }
 
-Vector2 Vector2::posmodv(const Vector2 &p_modv) const {
-	return Vector2(Math::fposmod(x, p_modv.x), Math::fposmod(y, p_modv.y));
+real_t Vector2::angle_to_point(const Vector2 &p_vector2) const {
+	return (p_vector2 - *this).angle();
 }
+
+Vector2 Vector2::from_angle(real_t p_angle) {
+	return Vector2(Math::cos(p_angle), Math::sin(p_angle));
+}
+
+
+real_t Vector2::distance_to(const Vector2 &p_vector2) const {
+	return Math::sqrt((x - p_vector2.x) * (x - p_vector2.x) + (y - p_vector2.y) * (y - p_vector2.y));
+}
+
+real_t Vector2::distance_squared_to(const Vector2 &p_vector2) const {
+	return (x - p_vector2.x) * (x - p_vector2.x) + (y - p_vector2.y) * (y - p_vector2.y);
+}
+
+_FORCE_INLINE_ Vector2 Vector2::direction_to(const Vector2 &p_to) const {
+	Vector2 ret(p_to.x - x, p_to.y - y);
+	ret.normalize();
+	return ret;
+}
+
 
 Vector2 Vector2::project(const Vector2 &p_to) const {
 	return p_to * (dot(p_to) / p_to.length_squared());
 }
 
-Vector2 Vector2::clamp(const Vector2 &p_min, const Vector2 &p_max) const {
-	return Vector2(
-			CLAMP(x, p_min.x, p_max.x),
-			CLAMP(y, p_min.y, p_max.y));
+Vector2 Vector2::plane_project(real_t p_d, const Vector2 &p_vec) const {
+	return p_vec - *this * (dot(p_vec) - p_d);
 }
 
-Vector2 Vector2::clampf(real_t p_min, real_t p_max) const {
-	return Vector2(
-			CLAMP(x, p_min, p_max),
-			CLAMP(y, p_min, p_max));
+
+real_t Vector2::dot(const Vector2 &p_other) const {
+	return x * p_other.x + y * p_other.y;
 }
 
-Vector2 Vector2::snapped(const Vector2 &p_step) const {
-	return Vector2(
-			Math::snapped(x, p_step.x),
-			Math::snapped(y, p_step.y));
+real_t Vector2::cross(const Vector2 &p_other) const {
+	return x * p_other.y - y * p_other.x;
 }
 
-Vector2 Vector2::snappedf(real_t p_step) const {
-	return Vector2(
-			Math::snapped(x, p_step),
-			Math::snapped(y, p_step));
+Vector2 Vector2::orthogonal() const {
+	return Vector2(y, -x);
 }
 
-Vector2 Vector2::limit_length(real_t p_len) const {
-	const real_t l = length();
-	Vector2 v = *this;
-	if (l > 0 && p_len < l) {
-		v /= l;
-		v *= p_len;
+
+_FORCE_INLINE_ Vector2::Axis Vector2::min_axis_index() const {
+	return x < y ? Vector2::AXIS_X : Vector2::AXIS_Y;
+}
+_FORCE_INLINE_ Vector2::Axis Vector2::max_axis_index() const {
+	return x < y ? Vector2::AXIS_Y : Vector2::AXIS_X;
+}
+
+
+Vector2 Vector2::lerp(const Vector2 &p_to, real_t p_weight) const {
+	Vector2 res = *this;
+	res.x = Math::lerp(res.x, p_to.x, p_weight);
+	res.y = Math::lerp(res.y, p_to.y, p_weight);
+	return res;
+}
+
+Vector2 Vector2::slerp(const Vector2 &p_to, real_t p_weight) const {
+	real_t start_length_sq = length_squared();
+	real_t end_length_sq = p_to.length_squared();
+	if (unlikely(start_length_sq == 0.0f || end_length_sq == 0.0f)) {
+		// Zero length vectors have no angle, so the best we can do is either lerp or throw an error.
+		return lerp(p_to, p_weight);
 	}
-
-	return v;
+	real_t start_length = Math::sqrt(start_length_sq);
+	real_t result_length = Math::lerp(start_length, Math::sqrt(end_length_sq), p_weight);
+	real_t angle = angle_to(p_to);
+	return rotated(angle * p_weight) * (result_length / start_length);
 }
 
-Vector2 Vector2::move_toward(const Vector2 &p_to, real_t p_delta) const {
-	Vector2 v = *this;
-	Vector2 vd = p_to - v;
-	real_t len = vd.length();
-	return len <= p_delta || len < (real_t)CMP_EPSILON ? p_to : v + vd / len * p_delta;
+Vector2 Vector2::cubic_interpolate(const Vector2 &p_b, const Vector2 &p_pre_a, const Vector2 &p_post_b, real_t p_weight) const {
+	Vector2 res = *this;
+	res.x = Math::cubic_interpolate(res.x, p_b.x, p_pre_a.x, p_post_b.x, p_weight);
+	res.y = Math::cubic_interpolate(res.y, p_b.y, p_pre_a.y, p_post_b.y, p_weight);
+	return res;
 }
+
+Vector2 Vector2::cubic_interpolate_in_time(const Vector2 &p_b, const Vector2 &p_pre_a, const Vector2 &p_post_b, real_t p_weight, real_t p_b_t, real_t p_pre_a_t, real_t p_post_b_t) const {
+	Vector2 res = *this;
+	res.x = Math::cubic_interpolate_in_time(res.x, p_b.x, p_pre_a.x, p_post_b.x, p_weight, p_b_t, p_pre_a_t, p_post_b_t);
+	res.y = Math::cubic_interpolate_in_time(res.y, p_b.y, p_pre_a.y, p_post_b.y, p_weight, p_b_t, p_pre_a_t, p_post_b_t);
+	return res;
+}
+
+Vector2 Vector2::bezier_interpolate(const Vector2 &p_control_1, const Vector2 &p_control_2, const Vector2 &p_end, real_t p_t) const {
+	Vector2 res = *this;
+	res.x = Math::bezier_interpolate(res.x, p_control_1.x, p_control_2.x, p_end.x, p_t);
+	res.y = Math::bezier_interpolate(res.y, p_control_1.y, p_control_2.y, p_end.y, p_t);
+	return res;
+}
+
+Vector2 Vector2::bezier_derivative(const Vector2 &p_control_1, const Vector2 &p_control_2, const Vector2 &p_end, real_t p_t) const {
+	Vector2 res = *this;
+	res.x = Math::bezier_derivative(res.x, p_control_1.x, p_control_2.x, p_end.x, p_t);
+	res.y = Math::bezier_derivative(res.y, p_control_1.y, p_control_2.y, p_end.y, p_t);
+	return res;
+}
+
 
 // slide returns the component of the vector along the given plane, specified by its normal vector.
 Vector2 Vector2::slide(const Vector2 &p_normal) const {
@@ -192,6 +241,64 @@ Vector2 Vector2::reflect(const Vector2 &p_normal) const {
 	return 2.0f * p_normal * dot(p_normal) - *this;
 }
 
+Vector2 Vector2::move_toward(const Vector2 &p_to, real_t p_delta) const {
+	Vector2 v = *this;
+	Vector2 vd = p_to - v;
+	real_t len = vd.length();
+	return len <= p_delta || len < (real_t)CMP_EPSILON ? p_to : v + vd / len * p_delta;
+}
+
+
+Vector2 Vector2::posmod(real_t p_mod) const {
+	return Vector2(Math::fposmod(x, p_mod), Math::fposmod(y, p_mod));
+}
+
+Vector2 Vector2::posmodv(const Vector2 &p_modv) const {
+	return Vector2(Math::fposmod(x, p_modv.x), Math::fposmod(y, p_modv.y));
+}
+
+
+Vector2 Vector2::min(const Vector2 &p_vector2) const {
+	return Vector2(MIN(x, p_vector2.x), MIN(y, p_vector2.y));
+}
+Vector2 Vector2::minf(real_t p_scalar) const {
+	return Vector2(MIN(x, p_scalar), MIN(y, p_scalar));
+}
+Vector2 Vector2::max(const Vector2 &p_vector2) const {
+	return Vector2(MAX(x, p_vector2.x), MAX(y, p_vector2.y));
+}
+Vector2 Vector2::maxf(real_t p_scalar) const {
+	return Vector2(MAX(x, p_scalar), MAX(y, p_scalar));
+}
+
+Vector2 Vector2::snapped(const Vector2 &p_step) const {
+	return Vector2(
+			Math::snapped(x, p_step.x),
+			Math::snapped(y, p_step.y));
+}
+
+Vector2 Vector2::snappedf(real_t p_step) const {
+	return Vector2(
+			Math::snapped(x, p_step),
+			Math::snapped(y, p_step));
+}
+
+Vector2 Vector2::clamp(const Vector2 &p_min, const Vector2 &p_max) const {
+	return Vector2(
+			CLAMP(x, p_min.x, p_max.x),
+			CLAMP(y, p_min.y, p_max.y));
+}
+
+Vector2 Vector2::clampf(real_t p_min, real_t p_max) const {
+	return Vector2(
+			CLAMP(x, p_min, p_max),
+			CLAMP(y, p_min, p_max));
+}
+
+
+real_t Vector2::aspect() const { return width / height; }
+
+
 bool Vector2::is_equal_approx(const Vector2 &p_v) const {
 	return Math::is_equal_approx(x, p_v.x) && Math::is_equal_approx(y, p_v.y);
 }
@@ -204,8 +311,85 @@ bool Vector2::is_zero_approx() const {
 	return Math::is_zero_approx(x) && Math::is_zero_approx(y);
 }
 
-bool Vector2::is_finite() const {
-	return Math::is_finite(x) && Math::is_finite(y);
+_FORCE_INLINE_ real_t &Vector2::operator[](int p_axis) {
+	DEV_ASSERT((unsigned int)p_axis < 2);
+	return coord[p_axis];
+}
+_FORCE_INLINE_ const real_t &Vector2::operator[](int p_axis) const {
+	DEV_ASSERT((unsigned int)p_axis < 2);
+	return coord[p_axis];
+}
+
+constexpr Vector2 Vector2::operator+(const Vector2 &p_v) const {
+	return Vector2(x + p_v.x, y + p_v.y);
+}
+constexpr void Vector2::operator+=(const Vector2 &p_v) {
+	x += p_v.x;
+	y += p_v.y;
+}
+
+constexpr Vector2 Vector2::operator-(const Vector2 &p_v) const {
+	return Vector2(x - p_v.x, y - p_v.y);
+}
+constexpr void Vector2::operator-=(const Vector2 &p_v) {
+	x -= p_v.x;
+	y -= p_v.y;
+}
+
+constexpr Vector2 Vector2::operator*(const Vector2 &p_v1) const {
+	return Vector2(x * p_v1.x, y * p_v1.y);
+}
+constexpr void Vector2::operator*=(const Vector2 &p_rvalue) {
+	*this = *this * p_rvalue;
+}
+
+constexpr Vector2 Vector2::operator/(const Vector2 &p_v1) const {
+	return Vector2(x / p_v1.x, y / p_v1.y);
+}
+constexpr void Vector2::operator/=(const Vector2 &p_rvalue) {
+	*this = *this / p_rvalue;
+}
+
+// Scalars
+
+constexpr Vector2 Vector2::operator*(real_t p_rvalue) const {
+	return Vector2(x * p_rvalue, y * p_rvalue);
+}
+constexpr void Vector2::operator*=(real_t p_rvalue) {
+	x *= p_rvalue;
+	y *= p_rvalue;
+}
+
+constexpr Vector2 Vector2::operator/(real_t p_rvalue) const {
+	return Vector2(x / p_rvalue, y / p_rvalue);
+}
+constexpr void Vector2::operator/=(real_t p_rvalue) {
+	x /= p_rvalue;
+	y /= p_rvalue;
+}
+
+constexpr Vector2 Vector2::operator-() const {
+	return Vector2(-x, -y);
+}
+
+// Comparison
+constexpr bool Vector2::operator==(const Vector2 &p_vec2) const {
+	return x == p_vec2.x && y == p_vec2.y;
+}
+constexpr bool Vector2::operator!=(const Vector2 &p_vec2) const {
+	return x != p_vec2.x || y != p_vec2.y;
+}
+constexpr bool Vector2::operator<(const Vector2 &p_vec2) const {
+	return x == p_vec2.x ? (y < p_vec2.y) : (x < p_vec2.x);
+}
+constexpr bool Vector2::operator>(const Vector2 &p_vec2) const {
+	return x == p_vec2.x ? (y > p_vec2.y) : (x > p_vec2.x);
+}
+constexpr bool Vector2::operator<=(const Vector2 &p_vec2) const {
+	return x == p_vec2.x ? (y <= p_vec2.y) : (x < p_vec2.x);
+}
+constexpr bool Vector2::operator>=(const Vector2 &p_vec2) const {
+	return x == p_vec2.x ? (y >= p_vec2.y) : (x > p_vec2.x);
 }
 
 Vector2::operator String() const {
@@ -214,4 +398,20 @@ Vector2::operator String() const {
 
 Vector2::operator Vector2i() const {
 	return Vector2i(x, y);
+}
+
+constexpr Vector2 operator*(float p_scalar, const Vector2 &p_vec) {
+	return p_vec * p_scalar;
+}
+
+constexpr Vector2 operator*(double p_scalar, const Vector2 &p_vec) {
+	return p_vec * p_scalar;
+}
+
+constexpr Vector2 operator*(int32_t p_scalar, const Vector2 &p_vec) {
+	return p_vec * p_scalar;
+}
+
+constexpr Vector2 operator*(int64_t p_scalar, const Vector2 &p_vec) {
+	return p_vec * p_scalar;
 }
