+++ core/math/vector2i.cpp
@@ -35,6 +35,50 @@
 #include "core/math/vector2.h"
 #include "core/string/ustring.h"
 
+
+double Vector2i::length() const {
+	return Math::sqrt((double)length_squared());
+}
+
+int64_t Vector2i::length_squared() const {
+	return x * (int64_t)x + y * (int64_t)y;
+}
+Vector2i Vector2i::sign() const { return Vector2i(SIGN(x), SIGN(y)); }
+Vector2i Vector2i::abs() const { return Vector2i(Math::abs(x), Math::abs(y)); }
+
+double Vector2i::distance_to(const Vector2i &p_to) const {
+	return (p_to - *this).length();
+}
+
+int64_t Vector2i::distance_squared_to(const Vector2i &p_to) const {
+	return (p_to - *this).length_squared();
+}
+
+
+_FORCE_INLINE_ Vector2i::Axis Vector2i::min_axis_index() const {
+	return x < y ? Vector2i::AXIS_X : Vector2i::AXIS_Y;
+}
+_FORCE_INLINE_ Vector2i::Axis Vector2i::max_axis_index() const {
+	return x < y ? Vector2i::AXIS_Y : Vector2i::AXIS_X;
+}
+
+
+Vector2i Vector2i::min(const Vector2i &p_vector2i) const {
+	return Vector2i(MIN(x, p_vector2i.x), MIN(y, p_vector2i.y));
+}
+
+Vector2i Vector2i::mini(int32_t p_scalar) const {
+	return Vector2i(MIN(x, p_scalar), MIN(y, p_scalar));
+}
+
+Vector2i Vector2i::max(const Vector2i &p_vector2i) const {
+	return Vector2i(MAX(x, p_vector2i.x), MAX(y, p_vector2i.y));
+}
+
+Vector2i Vector2i::maxi(int32_t p_scalar) const {
+	return Vector2i(MAX(x, p_scalar), MAX(y, p_scalar));
+}
+
 Vector2i Vector2i::clamp(const Vector2i &p_min, const Vector2i &p_max) const {
 	return Vector2i(
 			CLAMP(x, p_min.x, p_max.x),
@@ -59,12 +103,96 @@ Vector2i Vector2i::snappedi(int32_t p_step) const {
 			Math::snapped(y, p_step));
 }
 
-int64_t Vector2i::length_squared() const {
-	return x * (int64_t)x + y * (int64_t)y;
+real_t Vector2i::aspect() const { return width / (real_t)height; }
+
+
+_FORCE_INLINE_ int32_t &Vector2i::operator[](int p_axis) {
+	DEV_ASSERT((unsigned int)p_axis < 2);
+	return coord[p_axis];
+}
+_FORCE_INLINE_ const int32_t &Vector2i::operator[](int p_axis) const {
+	DEV_ASSERT((unsigned int)p_axis < 2);
+	return coord[p_axis];
 }
 
-double Vector2i::length() const {
-	return Math::sqrt((double)length_squared());
+// Vectors
+
+constexpr Vector2i Vector2i::operator+(const Vector2i &p_v) const {
+	return Vector2i(x + p_v.x, y + p_v.y);
+}
+constexpr void Vector2i::operator+=(const Vector2i &p_v) {
+	x += p_v.x;
+	y += p_v.y;
+}
+
+constexpr Vector2i Vector2i::operator-(const Vector2i &p_v) const {
+	return Vector2i(x - p_v.x, y - p_v.y);
+}
+constexpr void Vector2i::operator-=(const Vector2i &p_v) {
+	x -= p_v.x;
+	y -= p_v.y;
+}
+
+constexpr Vector2i Vector2i::operator*(const Vector2i &p_v1) const {
+	return Vector2i(x * p_v1.x, y * p_v1.y);
+}
+
+constexpr Vector2i Vector2i::operator/(const Vector2i &p_v1) const {
+	return Vector2i(x / p_v1.x, y / p_v1.y);
+}
+
+constexpr Vector2i Vector2i::operator%(const Vector2i &p_v1) const {
+	return Vector2i(x % p_v1.x, y % p_v1.y);
+}
+
+// Scalar
+
+constexpr Vector2i Vector2i::operator*(int32_t p_rvalue) const {
+	return Vector2i(x * p_rvalue, y * p_rvalue);
+}
+constexpr void Vector2i::operator*=(int32_t p_rvalue) {
+	x *= p_rvalue;
+	y *= p_rvalue;
+}
+
+constexpr Vector2i Vector2i::operator/(int32_t p_rvalue) const {
+	return Vector2i(x / p_rvalue, y / p_rvalue);
+}
+constexpr void Vector2i::operator/=(int32_t p_rvalue) {
+	x /= p_rvalue;
+	y /= p_rvalue;
+}
+
+constexpr Vector2i Vector2i::operator%(int32_t p_rvalue) const {
+	return Vector2i(x % p_rvalue, y % p_rvalue);
+}
+constexpr void Vector2i::operator%=(int32_t p_rvalue) {
+	x %= p_rvalue;
+	y %= p_rvalue;
+}
+
+constexpr Vector2i Vector2i::operator-() const {
+	return Vector2i(-x, -y);
+}
+
+// Comparisons
+constexpr bool Vector2i::operator==(const Vector2i &p_vec2) const {
+	return x == p_vec2.x && y == p_vec2.y;
+}
+constexpr bool Vector2i::operator!=(const Vector2i &p_vec2) const {
+	return x != p_vec2.x || y != p_vec2.y;
+}
+constexpr bool Vector2i::operator<(const Vector2i &p_vec2) const {
+	return (x == p_vec2.x) ? (y < p_vec2.y) : (x < p_vec2.x);
+}
+constexpr bool Vector2i::operator>(const Vector2i &p_vec2) const {
+	return (x == p_vec2.x) ? (y > p_vec2.y) : (x > p_vec2.x);
+}
+constexpr bool Vector2i::operator<=(const Vector2i &p_vec2) const {
+	return x == p_vec2.x ? (y <= p_vec2.y) : (x < p_vec2.x);
+}
+constexpr bool Vector2i::operator>=(const Vector2i &p_vec2) const {
+	return x == p_vec2.x ? (y >= p_vec2.y) : (x > p_vec2.x);
 }
 
 Vector2i::operator String() const {
@@ -73,4 +201,21 @@ Vector2i::operator String() const {
 
 Vector2i::operator Vector2() const {
 	return Vector2((int32_t)x, (int32_t)y);
+}
+
+
+constexpr Vector2i operator*(int32_t p_scalar, const Vector2i &p_vector) {
+	return p_vector * p_scalar;
+}
+
+constexpr Vector2i operator*(int64_t p_scalar, const Vector2i &p_vector) {
+	return p_vector * p_scalar;
+}
+
+constexpr Vector2i operator*(float p_scalar, const Vector2i &p_vector) {
+	return p_vector * p_scalar;
+}
+
+constexpr Vector2i operator*(double p_scalar, const Vector2i &p_vector) {
+	return p_vector * p_scalar;
 }
